# infrastructure/swagger/openapi.yml
openapi: 3.0.3
info:
  title: KSA Islamic Financing Platform API
  description: 'Unified API documentation for the KSA Islamic Financing Platform.


    All requests go through Kong API Gateway.

    URL pattern: /service-name/api/v1/...'
  version: 1.0.0
  contact:
    name: KSA Financing Platform Team
servers:
- url: http://46.62.226.94:8000
  description: Kong API Gateway
...
paths:
...
  /customer-service/api/v1/reference-data/source-of-wealth:
    post:
      summary: '[SOW] Create Source of Wealth Option'
      tags:
      - Source of Wealth
      operationId: SOW__Create_Source_of_Wealth_Option
      description: 'Kong -> customer-service:8084

        JWT REQUIRED (admin/super_admin) — Creates a new source of wealth dropdown option.

        Code must be unique per tenant. Options are active by default.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                  example: FREELANCE_INCOME
                nameEn:
                  type: string
                  example: Freelance / Gig Income
                nameAr:
                  type: string
                  example: دخل العمل الحر
                descriptionEn:
                  type: string
                  example: Income from freelance or gig economy work
                descriptionAr:
                  type: string
                  example: الدخل من العمل الحر أو اقتصاد الأعمال المؤقتة
                displayOrder:
                  type: integer
                  example: 9
              required:
              - code
              - nameEn
              - nameAr
              - descriptionEn
              - descriptionAr
              - displayOrder
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request — validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized — invalid or missing JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Unprocessable entity — business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - BearerAuth: []
    get:
      summary: '[SOW] List All (Admin)'
      tags:
      - Source of Wealth
      operationId: SOW__List_All__Admin
      description: 'Kong -> customer-service:8084

        JWT REQUIRED (admin) — Lists ALL source of wealth options including inactive.

        For admin panel management view.'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request — validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized — invalid or missing JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Unprocessable entity — business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - BearerAuth: []
...
  /customer-service/api/v1/reference-data/source-of-funds:
    post:
      summary: '[SOF] Create Source of Funds Option'
      tags:
      - Source of Funds
      operationId: SOF__Create_Source_of_Funds_Option
      description: 'Kong -> customer-service:8084

        JWT REQUIRED (admin/super_admin) — Creates a new source of funds dropdown option.

        Code must be unique per tenant. Options are active by default.'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                  example: CRYPTO_PROCEEDS
                nameEn:
                  type: string
                  example: Cryptocurrency / Digital Asset Proceeds
                nameAr:
                  type: string
                  example: عائدات العملات الرقمية
                descriptionEn:
                  type: string
                  example: Proceeds from cryptocurrency or digital asset sales
                descriptionAr:
                  type: string
                  example: عائدات من بيع العملات الرقمية أو الأصول الرقمية
                displayOrder:
                  type: integer
                  example: 9
              required:
              - code
              - nameEn
              - nameAr
              - descriptionEn
              - descriptionAr
              - displayOrder
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request — validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized — invalid or missing JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Unprocessable entity — business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - BearerAuth: []
    get:
      summary: '[SOF] List All (Admin)'
      tags:
      - Source of Funds
      operationId: SOF__List_All__Admin
      description: 'Kong -> customer-service:8084

        JWT REQUIRED (admin) — Lists ALL source of funds options including inactive.

        For admin panel management view.'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request — validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized — invalid or missing JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Unprocessable entity — business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - BearerAuth: []
...
  /customer-service/api/v1/reference-data/net-worth-ranges:
    post:
      summary: '[NWR] Create Net Worth Range Option'
      tags:
      - Net Worth Ranges
      operationId: NWR__Create_Net_Worth_Range_Option
      description: 'Kong -> customer-service:8084

        JWT REQUIRED (admin/super_admin) — Creates a new net worth range dropdown option.

        Code must be unique per tenant. minValue and maxValue are in SAR.

        Use null for maxValue on open-ended ranges (e.g. ''Over 10M'').'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                code:
                  type: string
                  example: RANGE_10M_50M
                nameEn:
                  type: string
                  example: 10,000,000 - 50,000,000 SAR
                nameAr:
                  type: string
                  example: 10,000,000 - 50,000,000 ريال
                descriptionEn:
                  type: string
                  example: High net worth individual range
                descriptionAr:
                  type: string
                  example: نطاق الأفراد ذوي الثروات العالية
                minValue:
                  type: integer
                  example: 10000000
                maxValue:
                  type: integer
                  example: 50000000
                displayOrder:
                  type: integer
                  example: 8
              required:
              - code
              - nameEn
              - nameAr
              - descriptionEn
              - descriptionAr
              - minValue
              - maxValue
              - displayOrder
      responses:
        '201':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request — validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized — invalid or missing JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Unprocessable entity — business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - BearerAuth: []
    get:
      summary: '[NWR] List All (Admin)'
      tags:
      - Net Worth Ranges
      operationId: NWR__List_All__Admin
      description: 'Kong -> customer-service:8084

        JWT REQUIRED (admin) — Lists ALL net worth range options including inactive.

        For admin panel management view. Each option has minValue/maxValue in SAR.'
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '400':
          description: Bad request — validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized — invalid or missing JWT
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Resource not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Unprocessable entity — business rule violation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
      - BearerAuth: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token from /identity-service/api/v1/auth/login or /identity-service/api/v1/auth/login-with-pin
  schemas:
    ApiResponse:
      type: object
      description: Standard success response envelope (from foundational-infra-sdk ApiResponse<T>)
      properties:
        data:
          description: Response payload — type varies per endpoint
        message:
          type: string
          example: OK
          description: Human-readable status message
        timestamp:
          type: string
          format: date-time
          example: '2026-02-26T07:36:23.118Z'
          description: ISO-8601 response timestamp
      required:
      - data
      - message
      - timestamp
    ErrorResponse:
      type: object
      description: Standard error response envelope (from foundational-infra-sdk ErrorResponse)
      properties:
        timestamp:
          type: string
          format: date-time
          example: '2026-02-26T07:36:23Z'
          description: ISO-8601 error timestamp
        error_code:
          type: integer
          example: 422
          description: HTTP status code
        error:
          type: string
          example: Unprocessable Entity
          description: HTTP status reason phrase
        code:
          type: string
          example: CUSTOMER.PROFILE.NOT_FOUND
          description: Domain error code (dot-separated)
        message:
          type: string
          example: Customer profile not found
          description: 'Localized error message (Accept-Language: en/ar)'
        path:
          type: string
          example: /api/v1/customers/123
          description: Request path that caused the error
        traceId:
          type: string
          example: a1b2c3d4
          description: Distributed tracing ID for debugging
        details:
          type: object
          additionalProperties:
            type: string
          description: Field-level validation errors (only for 400 responses)
          example:
            nationalId: must not be blank
            pin: size must be between 6 and 6
      required:
      - timestamp
      - error_code
      - error
      - code
      - message
      - path